% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mcmc_example}
\alias{mcmc_example}
\title{Simple mcmc example}
\usage{
mcmc_example(x0, w, nsteps, pars)
}
\arguments{
\item{x0}{Initial point}

\item{w}{Propopsal widths (uniform distribution)}

\item{nsteps}{Number of steps}

\item{pars}{Parameters (means and standard deviations)}
}
\description{
Simple mcmc example.  In this model the target distribution is an
n dimensional multivariate normal with zero correlations.  The
sampler will take a step from a uniform distribution with a width
\code{w[i] * 2}.
}

